Add --events flag to mysqldump command. Skip performance_schema db.
MySQL has recently added a warning to mysqldump output to alert users that Events are not dumped by default (see MySQL bug #55587). This warning causes spurious emails from cron. It doesn't hurt to back up Events so we add the --events flag to avoid the warning. Unfortunately this causes permission errors with the performance_schema db, even though mysqldump should ignore that db. Hence, we explicitely remove it from the to do list.