Allow skipping of deletion phase. when SKIP_DELETE is to true the script only writes new backups and creates new buckets. [SYSPRJ-36]
No attempts are made to delete data from S3. Set this to true when the credentials you provide don't have the rights to perform deletions. This is a handy way to prevent disaster should someone malicious gain access to them. With correctly restricted IAM permissions they won't be able to delete existing backups. You'll need to use a separate server and credentials to purge old backups.